+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Full documentation for RPP is available at [https://rocm.docs.amd.com/projects/r
 * Bitwise NOT, Bitwise AND, Bitwise OR augmentations on HOST (CPU) and HIP backends. (#520)
 * Tensor Concat augmentation on HOST (CPU) and HIP backends. (#530)
 * JPEG Compression Distortion augmentation on HIP backend. (#538)
+* `log1p`, defined as `log (1 + x)`, tensor augmentation support on HOST (CPU) and HIP backends.
 
 ### Changed